feat: filename fixes and fixed bug when user routing back to the result page and seeing loading skeleton instead of audio list
