--- src/components/header/api-link/api-link.tsx
@@ -1,11 +1,0 @@
-import style from './style.module.css';
-
-function ApiLink() {
-  return (
-    <div className={style.container}>
-      <a href="https://rapidapi.com/deezerdevs/api/deezer-1">API</a>
-    </div>
-  );
-}
-
-export default ApiLink;
