--- src/components/header/api-link/style.module.css
@@ -1,46 +1,0 @@
-@keyframes borderBlinking {
-  0% {
-    border-bottom-color: rgba(157, 154, 68, 1);
-    border-right-color: rgba(157, 154, 68, 1);
-  }
-  40% {
-    border-bottom-color: rgba(157, 154, 68, 1);
-    border-right-color: rgba(157, 154, 68, 1);
-  }
-  45% {
-    border-bottom-color: rgba(157, 154, 68, .2);
-  }
-  50% {
-    border-bottom-color: rgba(157, 154, 68, 1);
-    border-right-color: rgba(157, 154, 68, 1);
-  }
-  52% {
-    border-bottom-color: rgba(157, 154, 68, .6);
-    border-right-color: rgba(157, 154, 68, 1);
-  }
-  70% {
-    border-bottom-color: rgba(157, 154, 68, 1);
-    border-right-color: rgba(157, 154, 68, 1);
-  }
-  72% {
-    border-bottom-color: rgba(157, 154, 68, 1);
-    border-right-color: rgba(157, 154, 68, .2);
-  }
-  76% {
-    border-bottom-color: rgba(157, 154, 68, 1);
-  }
-  100% {
-    border-bottom-color: rgba(157, 154, 68, 1);
-    border-right-color: rgba(157, 154, 68, 1);
-  }
-}
-
-.container {
-  padding: 10px;
-  text-align: center;
-  transform: skewY(-10deg);
-  animation: borderBlinking 4s ease-in infinite alternate;
-  border: 3px solid rgba(157, 154, 68, 1);
-  border-radius: 6px;
-  box-shadow: 1px -1px 10px 5px rgba(157, 154, 68, .2);
-}
