+++ src/components/header/navigation-list/data.ts
@@ -1,0 +1,12 @@
+import FavoriteIcon from "@src/icons/favorite-icon.tsx";
+
+import { ROUTES } from "@src/const/routes.ts";
+
+const data = [
+  {
+    icon: FavoriteIcon,
+    href: ROUTES.FAVORITES,
+  },
+];
+
+export default data;
