+++ src/const/common.ts
@@ -2,5 +2,4 @@ export const HUNDRED_PERCENT = 100;
 
 export const DEFAULT_AUDIO_STATE = { time: 0, duration: 0 };
 
-export const DATA_ON_PAGE = 25;
 export const FAVORITE_ON_PAGE = 5;
