+++ src/icons/home-icon.tsx
@@ -1,0 +1,14 @@
+import { TDefaultIconProps } from "@src/icons/type.ts";
+
+function HomeIcon({ className = "" }: TDefaultIconProps) {
+  return (
+    <svg className={className} width="100%" viewBox="0 0 24 24" version="1.1" xmlSpace="preserve" xmlns="http://www.w3.org/2000/svg" xmlnsXlink="http://www.w3.org/1999/xlink">
+      <g id="grid_system"/>
+      <g id="_icons">
+        <path d="M7,21h10c2.2,0,4-1.8,4-4v-6.5c0-1.3-0.6-2.4-1.6-3.2l-5-3.8C13,2.5,11,2.5,9.6,3.6l-5,3.7C3.6,8.1,3,9.2,3,10.5V17   C3,19.2,4.8,21,7,21z M5,10.5c0-0.6,0.3-1.2,0.8-1.6l5-3.8c0.4-0.3,0.8-0.4,1.2-0.4s0.8,0.1,1.2,0.4l5,3.8c0.5,0.4,0.8,1,0.8,1.6   V17c0,1.1-0.9,2-2,2H7c-1.1,0-2-0.9-2-2V10.5z"/>
+      </g>
+    </svg>
+  );
+}
+
+export default HomeIcon;
