+++ src/icons/type.ts
@@ -1,0 +1,3 @@
+export type TDefaultIconProps = {
+  className?: string,
+}
