+++ src/pages/favorite-page/empty-screen/empty-screen.tsx
@@ -10,7 +10,7 @@ function EmptyScreen() {
           <img src="public/images/empty-favorite.png" alt="Empty favorite img illustration" />
         </div>
         <figcaption className="mt-4 text-center">
-          <p>Favorite list is empty, <Link to={ROUTES.INDEX} className="underline">go back</Link></p>
+          <p>Favorite list is empty, <Link to={ROUTES.INDEX} className="underline hover:opacity-80 active:opacity-60">go back</Link></p>
         </figcaption>
       </figure>
     </div>
