+++ src/pages/result-page/music-player/footer/audio-line/audio-line.tsx
