+++ src/pages/result-page/music-player/footer/audio-line/style.module.css
