+++ src/pages/result-page/music-player/footer/time-info/time-info.tsx
@@ -1,0 +1,13 @@
+import useZustandStore from "@zustand/zustandStore.ts";
+
+function TimeInfo() {
+  const currentAudioInfo = useZustandStore(state => state.currentAudioInfo);
+
+  return (
+    <div>
+      <p className="text-sm">{Math.floor(currentAudioInfo.time)}s/{Math.floor(currentAudioInfo.duration)}s</p>
+    </div>
+  );
+}
+
+export default TimeInfo;
