+++ src/pages/result-page/result-page.tsx
@@ -4,8 +4,8 @@ import useSearch from "@src/api/swr-api/useSearch/useSearch.ts";
 
 import useZustandStore from "@zustand/zustandStore.ts";
 
-import AudioCard from "@src/pages/result-page/audio-card/audio-card.tsx";
-import AudioSidebar from "./audio-sidebar/audio-sidebar.tsx";
+import MusicPlayer from "@result-page/music-player/music-player.tsx";
+import AudioSidebar from "@result-page/sidebar/audio-sidebar.tsx";
 import LoadingScreen from "./loading-screen/loading-screen.tsx";
 
 import { ISearch } from "@src/api/interfaces.ts";
@@ -15,20 +15,28 @@ function ResultPage() {
 
   const currentAudioId = useZustandStore(state => state.currentAudioId);
   const changeCurrentAudio = useZustandStore(state => state.changeCurrentAudio);
+  const resetAllAudioInfo = useZustandStore(state => state.resetAllAudioInfo);
+
+  const currentAudio = audioData?.find((item) => Number(item.id) === currentAudioId);
+
+  useEffect(() => {
+    if (audioData && !currentAudio) {
+      changeCurrentAudio(Number(audioData[0].id));
+    }
+  }, [audioData]);
 
   useEffect(() => {
     return () => {
       changeCurrentAudio(null);
+      resetAllAudioInfo();
     }
-  }, []);
-
-  const currentAudio = audioData?.find((item) => Number(item.id) === currentAudioId);
+  }, [])
 
   return (
     <section>
       {currentAudio ? (
         <div className="mx-auto w-full lg:w-10/12 xl:w-8/12 flex flex-col lg:flex-row justify-between">
-          <AudioCard audioData={audioData as ISearch[]} currentAudio={currentAudio} />
+          <MusicPlayer audioData={audioData as ISearch[]} currentAudio={currentAudio} />
           <AudioSidebar />
         </div>
       ) : <LoadingScreen isLoading={isLoading} isError={Boolean(isError)} audioData={audioData} mutate={mutate} />}
