+++ src/zustand/slices/audio/createAudioInfo.ts
@@ -7,6 +7,7 @@ export interface IAudioInfo {
   updateCurrentAudioInfo: (currentTime: number, duration?: number) => void,
   currentAudioPlayState: boolean,
   updateCurrentAudioPlayState: (newPlayState?: boolean) => void,
+  resetAllAudioInfo: () => void,
 }
 
 const createAudioInfo: StateCreator<IAudioInfo, [["zustand/devtools", never]]> = (set, get) => ({
@@ -14,6 +15,7 @@ const createAudioInfo: StateCreator<IAudioInfo, [["zustand/devtools", never]]> =
   updateCurrentAudioInfo: (time: number, duration?: number) => set(({ currentAudioInfo }) => ({ currentAudioInfo: { time, duration: duration || currentAudioInfo.duration }  }), false, "updateCurrentAudioInfo"),
   currentAudioPlayState: false,
   updateCurrentAudioPlayState: (newPlayState?: boolean) => set({ currentAudioPlayState: newPlayState !== undefined ? newPlayState : !get().currentAudioPlayState }, false, "updateCurrentAudioPlayState"),
+  resetAllAudioInfo: () => set({ currentAudioInfo: DEFAULT_AUDIO_STATE, currentAudioPlayState: false }, false, "resetAllAudioInfo"),
 })
 
 export default createAudioInfo;
